+++ docs/.vuepress/config.js
@@ -145,6 +145,7 @@ module.exports = {
           children: [
             'task_library/',
             'task_library/airflow',
+            'task_library/airtable',
             'task_library/aws',
             'task_library/collections',
             'task_library/constants',
@@ -158,6 +159,7 @@ module.exports = {
             'task_library/shell',
             'task_library/sqlite',
             'task_library/strings',
+            'task_library/twitter',
           ]
         },
         {
@@ -205,6 +207,7 @@ module.exports = {
             'examples/map_reduce',
             'examples/parameterized_flow',
             'examples/retries_with_mapping',
+            'examples/twitter_to_airtable',
           ]
         },
         {