+++ src/prefect/tasks/airtable/__init__.py
@@ -1,0 +1,6 @@
+try:
+    from prefect.tasks.airtable.airtable import WriteAirtableRow
+except ImportError:
+    raise ImportError(
+        'Using `prefect.tasks.airtable` requires Prefect to be installed with the "airtable" extra.'
+    )